Retargetted SpecsFor<Web> Helpers to .NET 4.5, bumped version to 1.4.0.

+++ SpecsFor.Helpers.Web/PublishPackage.ps1
@@ -1,9 +1,29 @@
+Param(
+  [switch]$DoNotPush,
+  [switch]$PushToLocalFeed
+)
+
+$ProjectName = "SpecsFor.Helpers.Web.csproj"
+
+$LocalTestFeedDir = "C:\Projects\LocalNuGetFeed\"
+
 #Remove existing packages
 Remove-Item *.nupkg
-#Build release
 #Create package
-nuget pack SpecsFor.Helpers.Web.csproj -Build
-#Push
+nuget pack $ProjectName -Build -Properties Configuration=Debug
 $PackageName = gci "*.nupkg"
-nuget push $PackageName
-Remove-Item *.nupkg
+
+if ($PushToLocalFeed) {
+	Write-Host -ForegroundColor Yellow "Deploying to local NuGet test feed!"
+	Copy-Item $PackageName $LocalTestFeedDir
+}
+elseif ($DoNotPush) {
+	Write-Host -ForegroundColor Yellow "Skipping 'nuget push' step!"
+}
+else {
+	#Push
+	Write-Host "About to push to NuGet.org!  Press CTRL+C to abort."
+	pause
+	nuget push $PackageName
+	Remove-Item *.nupkg
+}
